Refactor chat template integration
