--- pkg/tokenization/chat_template_go/transformers.code-workspace
@@ -1,11 +1,0 @@
-{
-	"folders": [
-		{
-			"path": "../../../../transformers"
-		},
-		{
-			"path": "../../.."
-		}
-	],
-	"settings": {}
-}
