+++ requirements.txt
@@ -1,0 +1,5 @@
+packaging==24.2
+pillow==11.2.1
+torch==2.7.0
+transformers>=4.53.0
+jinja2>=2.11
